requirements.txt
discord.py
numpy
opencv-python
pytesseract
pyyaml